Major changes
Unnecessary lines removal in poseModule since we are not running the program in poseModule.py

Creating accuracy capturing function every 5 seconds and saving it into a list with random play_id and time.

After the game is over or quit, the accuracyList data is sent to the database into its corresponding table and columns.